+++ 00_signup/src/component/signUpForm/signUpInput.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react'
-import { user } from '../../type/userInfo'
+import { user } from '../../utils/userInfo.type'
 import '../../css/signUpInput.css'
 
 type ItemProps = {
