+++ 00_signup/src/utils/signUp.policy.ts
@@ -1,4 +1,4 @@
-import { user } from '../type/userInfo'
+import { user } from './userInfo.type'
 
 export const emailReg = RegExp(/(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))/);
 export const passwordReg = RegExp(/(?=.*[A-Za-z])(?=.*\d)(?=.*[$@$!%*#?&])[A-Za-z\d$@$!%*#?&]{8,42}/);
