+++ 00_signup/src/utils/userInfo.type.ts
